First step towards redundant mode
With this patch it's possible to run librist in a redundant setup, where one of the instances can fail without issue for the receiver.
To run redundant librist instances we need to run synchronized, the simplest way to do that is to sync from our input. This patch introduces an API call to change the flow_id if our source flow_id is to change.
Also it introduces a new option on rist_sender_data_write to tell librist to use the containing rtp_seq in the seq field of the rist_data_block

Todo: figure out how to loadbalance our outputs (i.e.: have instance 1 send odd sequenced packets, and instance 2 send even sequenced packets) without breaking receiver or triggering (too many) retransmits.