Add RTC delay based timing mode implementation
RTC timing mode is intended to add a delay to the RTC time as is transmitted via the RTCP SR packet. For this to work accurately the sender MUST set the correct NTP timestamp in it's SR packets and both sender & receiver clock MUST be synchronized via a timekeeping service like NTP or PTP.
This mode is useful for keeping streams synchronized across multiple receivers, or multiple streams within the same receiver. This mode calculates output time based on the NTP timestamp when a packet was send, and therefore ignores any network delays in between. Thus users must ensure they choose a big enough buffer size to account for this.