Fix cross build when using MbedTLS
Previous method was written only our MinGW setup in mind.
This corrects build failures due to missing cmake toolchainfile, which may not be needed for other builds (as tested with provided aarch64 cross file).
Meson >= 0.56.0 correctly generates the cmake toolchain file (and documentation explicitly states not to set this option manually), so there it is not used
The toolchain variable has been renamed to make more sense when used in other arches, and in a way that it should not clash with any meson native options.