Correctly set the version this time.
The RIST committee in all it's wisdom decided to use some of the reserved bits for the version, instead of the existing GRE version field (likely to keep somewhat of compatibility with existing GRE implementations).

"RFC 8086 defines bits 4-12 in the GRE header as Reserved0.  This Specification defines bits 9-12 as follows (see Figure 1):
o   RV (bits 10-12): this field indicates the RIST Version, as follows:
§  RV = 000: TR-06-2:2020
§  RV = 001: TR-06 -2-2021
o   H (bit 9): this field indicates the AES key length for PSK operation, as follows:
§  H = 0: 128-bit AES key
§  H = 1: 256-bit AES key
·         Usage of the H bit shall be as follows:
o   If RV = 000, the H bit shall be set to zero on transmission and ignored on reception.  In this case, the PSK key length shall be determined by out-of-band means.
o   If RV = 001 or higher, the H bit indicates the PSK key length.  Devices not operating in PSK mode shall set the H bit to zero on transmission and ignore it on reception.
·         The remaining Reserved0 bits in Figure 1 shall be set to zero on transmission and ignored on reception, as per RFC 2784."