Do not treat not having credentials while being requested for them as a permanent failure.
This fixes an issue whereby a sender is configured to connect outbound without credentials
while the receiver (in server mode) is configured to request them. This would put the sender
in a permanent failure state, where it cannot send data to the receiver, even if the receiver
configuration was changed to not require credentials.