Workaround/fixes for cJSON symbol collision
Provide 2 new confiration options "disable_json" and "external_cjson"
external_cjson is intended ONLY for static libraries that get linked
to an application that already has cjson compiled in, so that it may
use that. The version of cJSON MUST match the version used in libRIST
or at least be ABI compatible.