Fix building srp_examples.c with musl
This fixes the immediate build failure, but it's quite a fragile fix,
because if any other libc defines malloc, calloc, or free in a header
that has not been included by this point in srp_examples.c, but is
subsequently included by srp.c or srp_constants.c, the same problem
will occur.  The ideal thing to do would be to find a way to
accomplish the desired effect without needing to use a macro to
override the definitions of libc functions.